refactor: move schema metadata services from Data to Application plane
DataverseEntityMetadataService, DataverseRelationshipService, and
DataverseOptionSetService perform DDL-style schema operations (entity/
attribute/relationship create/update/delete) which belong in the
Application plane, not the Data plane.

The Data plane (Platform.Dataverse.Data) retains data-focused services:
DataverseFileService, DataverseRecordService, DataverseBulkService,
ChangesetApplier.
