+++ src/TALXIS.CLI.Platform.Dataverse.Application/Services/DataverseRelationshipService.cs
@@ -2,7 +2,7 @@ using Microsoft.Xrm.Sdk;
 using TALXIS.CLI.Core.Contracts.Dataverse;
 using TALXIS.CLI.Platform.Dataverse.Runtime;
 
-namespace TALXIS.CLI.Platform.Dataverse.Data;
+namespace TALXIS.CLI.Platform.Dataverse.Application.Services;
 
 /// <summary>
 /// Implements many-to-many (N:N) relationship associate/disassociate
